requirements.txt
streamlit
google-generativeai
pyMuPDF
python-dotenv
pillow
